refactor: extract installment parsing and expand plan types
Android:
- Create InstallmentConfigurationParser class following existing patterns
- Takes ReadableMap in constructor (like other parsers)
- Uses computed property instead of method
- Encapsulates installment constants and parsing logic
- CardConfigurationParser delegates to InstallmentConfigurationParser

TypeScript:
- Expand InstallmentPlan type to include all supported plan types:
  regular, revolving, bonus, with_interest, buynow_paylater,
  nointerest_bonus, interest_bonus, refund_prctg,
  nointerest_refund_prctg, interest_refund_prctg
